+++ NEWS.md
@@ -1,6 +1,7 @@
 # statgenHTP 1.0.1.1
 
 * Plotting second derivatives of fitted splines, plotted first derivatives. This is fixed now. 
+* The detectSerieOut function now uses an extra criterion for checking if time courses our outlying. See the function documentation and vignettes for a full explanation of this new criterion.
 
 # statgenHTP 1.0.1