+++ NEWS.md
@@ -1,7 +1,8 @@
 # statgenHTP 1.0.1.1
 
 * Plotting second derivatives of fitted splines, plotted first derivatives. This is fixed now. 
-* The detectSerieOut function now uses an extra criterion for checking if time courses our outlying. See the function documentation and vignettes for a full explanation of this new criterion.
+* The ```detectSerieOut```` function now uses an extra criterion for checking if time courses our outlying. See the function documentation and vignettes for a full explanation of this new criterion.
+* The parameter ```trait``` in the ```removeSerieOut``` function is renamed to ```traits``` and now accepts a vector of traits that for which outlier values can be replaced by ```NA```.
 
 # statgenHTP 1.0.1