+++ NEWS.md
@@ -1,4 +1,8 @@
-# statgenHTP 1.0.1.1
+# statgenHTP 1.0.3
+
+* No user visible changes
+
+# statgenHTP 1.0.2
 
 * Plotting second derivatives of fitted splines, plotted first derivatives. This is fixed now. 
 * The ```detectSerieOut```` function now uses an extra criterion for checking if time courses our outlying. See the function documentation and vignettes for a full explanation of this new criterion.